Use provider cache to use the same result for capability instead of multiple requests
