+++ routes/ai-home/components/DeveloperSettings.tsx
@@ -1,34 +1,17 @@
 /**
  * WordPress dependencies
  */
-import apiFetch from '@wordpress/api-fetch';
 import { Button, Spinner } from '@wordpress/components';
 import { DataForm } from '@wordpress/dataviews';
 import type { Field, Form } from '@wordpress/dataviews';
-import {
-	useCallback,
-	useEffect,
-	useMemo,
-	useRef,
-	useState,
-} from '@wordpress/element';
+import { useCallback, useMemo, useRef } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 
 /**
  * Internal dependencies
  */
 import { useDeveloperFeatureSettings } from '../hooks/use-developer-feature-settings';
-
-interface ModelData {
-	id: string;
-	name: string;
-}
-
-interface ProviderData {
-	id: string;
-	name: string;
-	models: ModelData[];
-}
+import { useProviders } from '../hooks/use-providers';
 
 interface DeveloperSettingsProps {
 	featureId: string;
@@ -55,41 +38,12 @@ export function DeveloperSettings( {
 	featureId,
 	capability,
 }: DeveloperSettingsProps ): React.JSX.Element {
-	const [ providers, setProviders ] = useState< ProviderData[] >( [] );
-	const [ isLoading, setIsLoading ] = useState( true );
-	const [ fetchError, setFetchError ] = useState< string | null >( null );
+	const { providers, isLoading, fetchError } = useProviders( capability );
 
 	const formWrapperRef = useRef< HTMLDivElement >( null );
 
 	const { settings, update, clear, isSaving } =
 		useDeveloperFeatureSettings( featureId );
-
-	useEffect( () => {
-		if ( capability === 'none' ) {
-			setProviders( [] );
-			setFetchError( null );
-			setIsLoading( false );
-			return;
-		}
-
-		setIsLoading( true );
-		setFetchError( null );
-
-		apiFetch< ProviderData[] >( {
-			path: `/ai/v1/providers?capability=${ encodeURIComponent(
-				capability
-			) }`,
-		} )
-			.then( ( data ) => {
-				setProviders( data );
-			} )
-			.catch( () => {
-				setFetchError( __( 'Failed to load providers.', 'ai' ) );
-			} )
-			.finally( () => {
-				setIsLoading( false );
-			} );
-	}, [ capability ] );
 
 	const getModelElements = useCallback( () => {
 		const provider = providers.find( ( p ) => p.id === settings.provider );
